+++ js/auditRecords/index.js
@@ -1,2 +1,1 @@
-export * from './AuditRecordsAPI'
 export * from './AuditRecords'
